+++ rts/motoko-rts/src/gc/mark_compact.rs
@@ -7,7 +7,7 @@ use bitmap::{alloc_bitmap, free_bitmap, get_bit, iter_bits, set_bit, BITMAP_ITER
 use mark_stack::{alloc_mark_stack, free_mark_stack, pop_mark_stack};
 
 use crate::heap::Heap;
-use crate::mem::memcpy_words;
+use crate::mem_utils::memcpy_words;
 use crate::types::*;
 use crate::visitor::visit_pointer_fields;
