+++ rts/motoko-rts/src/mem_utils.rs
